Refactor `Gettext` by creating `Gettext.Compiler`
I moved *all* of the compilation logic (which transformed `.po` files into
function clauses) used in the `Gettext` module to the `Gettext.Compiler` module.
That's it.
